requirements.txt
amqp==5.2.0
asgiref==3.8.1
bcrypt==4.1.3
billiard==4.2.0
boto3==1.35.4
botocore==1.35.4
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
diff-match-patch==20230430
Django==5.0.6
django-admin-interface==0.28.8
django-autoslug==1.9.9
django-cities==0.6.1
django-cities-light==3.10.1
django-colorfield==0.11.0
django-import-export==4.1.1
django-jazzmin==3.0.0
django-json-widget==2.0.1
django-model-utils==4.5.1
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et_xmlfile==2.0.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
jmespath==1.0.1
jwt==1.3.1
kombu==5.3.7
numpy==2.0.1
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.0
password==0.2
pillow==10.3.0
progressbar2==4.4.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.11
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
rfc3986==1.5.0
s3transfer==0.10.2
setuptools==80.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
swapper==1.3.0
tablib==3.5.0
text-unidecode==1.3
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
